requirements.txt
appnope==0.1.2
argcomplete==2.0.0
asgiref==3.5.0
astroid==2.9.3
asttokens==2.0.5
attrs==21.4.0
Authlib==1.0.1
autopep8==1.6.0
backcall==0.2.0
binarytree==6.3.0
cachetools==5.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
coverage==6.2
cryptography==37.0.2
decorator==5.1.1
cssselect==1.1.0
defusedxml==0.7.1
distlib==0.3.4
dj-database-url==0.5.0
dj-rest-auth==2.2.4
Django==4.0.4
django-allauth==0.51.0
django-cors-headers==3.12.0
django-extensions==3.1.5
django-filter==22.1
django-heroku==0.3.1
django-phonenumber-field==6.3.0
django-rest-framework==0.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
ecdsa==0.17.0
executing==0.8.3
filelock==3.7.0
Flask==2.1.2
Flask-Cors==3.0.10
graphviz==0.19.1
google-api-core==2.8.2
google-api-python-client==2.52.0
google-auth==2.8.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.3
green==3.4.0
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
importlib-metadata==4.11.4
iniconfig==1.1.1
ipython==8.1.1
isort==5.10.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.0.3
lazy-object-proxy==1.7.1
lxml==4.7.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mccabe==0.6.1
money==1.3.0
npm==0.1.1
numpy==1.22.2
oauthlib==3.2.0
optional-django==0.1.0
packaging==21.3
pandas==1.4.1
parso==0.8.3
pep8==1.7.1
pexpect==4.8.0
phonenumbers==8.12.50
pickleshare==0.7.5
Pillow==9.1.1
pipenv==2022.5.2
pipx==1.0.0
platformdirs==2.5.0
pluggy==1.0.0
prompt-toolkit==3.0.28
protobuf==4.21.2
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.12.0
PyJWT==2.4.0
pylint==2.12.2
pytest==7.0.0
pyparsing==3.0.9
pyquery==1.4.3
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.19.2
python-jose==3.3.0
python3-openid==3.2.0
pytz==2021.3
redgreenunittest==0.1.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
setuptools-scm==6.4.2
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sportsipy==0.6.0
sqlparse==0.4.2
stack-data==0.2.0
toml==0.10.2
tomli==2.0.0
traitlets==5.1.1
twilio==7.9.3
typing_extensions==4.1.1
Unidecode==1.3.2
uritemplate==4.1.1
urllib3==1.26.9
userpath==1.8.0
virtualenv==20.14.1
virtualenv-clone==0.5.7
wcwidth==0.2.5
Werkzeug==2.1.2
whitenoise==6.2.0
wrapt==1.13.3
xmltodict==0.13.0
zipp==3.8.0
